refactor(delegation): collapse the shallow delivery-format files (A4)
Merge subagentDeliveryFormat.ts and its single-caller subagentDiffs.ts
leaf into subagentResults.ts, inlining the formatBuiltSubagentDelivery
arg-rearranger at its one call site. Dissolve the follow-up queue
manager's drainItems/restore one-line delegations — callers go through
the already-guarded queue(lease) accessor. A4 of the delegation/flow
substrate consolidation proposal (childRunDelivery.ts and the
workflowScriptAgentRunner/createRunAgent seam stay, per the verifier
rulings).
