+++ packages/cli/src/chat/chatSessionController.ts
@@ -257,7 +257,7 @@ export function createChatSessionController(
     recovery: SupersededInterruptedRecovery | undefined,
     lease: FollowUpRecoveryLease,
   ): void => {
-    runtimeSession.followUps.restore(lease, recovery?.followUps ?? []);
+    runtimeSession.followUps.queue(lease).restore(recovery?.followUps ?? []);
     if (recovery?.followUps.length) {
       runtimeSession.events.emit({
         scope: 'session',
