+++ src/test-kernel/agent/followUp/FollowUpQueue.vitest.ts
@@ -97,7 +97,7 @@ describe('ToolUseFollowUpQueue ownership', () => {
 
     expect(queues.release(child, 'terminal')).toBe(false);
     expect(queues.getAll(id)).toEqual(['before handoff', 'during recovery']);
-    expect(queues.drainItems(recovery).map((item) => item.text)).toEqual([
+    expect(queues.queue(recovery).drainItems().map((item) => item.text)).toEqual([
       'before handoff',
       'during recovery',
     ]);
@@ -132,8 +132,8 @@ describe('ToolUseFollowUpQueue ownership', () => {
     a.queue(aLease).enqueue({ text: 'a' });
     b.queue(bLease).enqueue({ text: 'b' });
 
-    expect(a.drainItems(aLease).map((item) => item.text)).toEqual(['a']);
-    expect(b.drainItems(bLease).map((item) => item.text)).toEqual(['b']);
+    expect(a.queue(aLease).drainItems().map((item) => item.text)).toEqual(['a']);
+    expect(b.queue(bLease).drainItems().map((item) => item.text)).toEqual(['b']);
   });
 
   it('never reopens a terminal stream', () => {
@@ -228,7 +228,7 @@ describe('ToolUseFollowUpQueue ownership', () => {
         'persisted-generation',
       ),
     ).toEqual({ kind: 'unavailable' });
-    expect(queues.drainItems(live)).toEqual([]);
+    expect(queues.queue(live).drainItems()).toEqual([]);
   });
 
   it('does not reopen a terminal stream for an unrelated execution', () => {
@@ -276,7 +276,7 @@ describe('ToolUseFollowUpQueue ownership', () => {
     ).toBeDefined();
     expect(recovery.generationId).toBe('persisted-generation');
     expect(queues.currentGenerationId(id)).toBe('persisted-generation');
-    expect(queues.drainItems(recovery).map((item) => item.text)).toEqual([
+    expect(queues.queue(recovery).drainItems().map((item) => item.text)).toEqual([
       'continue',
     ]);
   });
