+++ src/test-kernel/tools/SubagentDeliveryStructured.vitest.ts
@@ -4,7 +4,7 @@ import { describe, expect, it } from 'vitest';
 // Local imports
 import type { AgentFlowResult } from '@agent/runtime/AgentFlowResult';
 import type { ExecutionId, StreamTabId } from '@shared/schemas';
-import { buildSubagentResult } from '@tools/delegation/subagentDeliveryFormat';
+import { buildSubagentResult } from '@tools/delegation/subagentResults';
 
 // A tool-use flow result carrying a value captured by the submit_output
 // terminal tool must surface as `.structured` on the built AgentFinalResult,
